Add Win32Exception.Create helpers and Madowaku.Io exception hierarchy
- Win32ExceptionExtensions.Create(WIN32_ERROR/HRESULT) sets both HResult and NativeErrorCode correctly (public setter on net10, cached reflection on net472/net481).
- Add Madowaku.Io exception hierarchy (MadowakuIoException, DriveLockedException, DriveNotReadyException, FileExistsException) with correct HResult.
- Error.cs: use Win32Exception.Create, modernize WindowsErrorToException to a switch expression, add ReadOnlySpan<string> FormatMessage overload.
- StringParameterArray: add ReadOnlySpan<string> params overload.
- Strip stray BOM from ComScopeTests.cs and VariantTests.cs; add test coverage for all new types.

+++ madowaku.tests/Windows/Win32/Foundation/HGLOBALTests.cs
@@ -17,7 +17,7 @@ public unsafe class HGLOBALTests
 
         try
         {
-            global.Size.Should().Be((nuint)(sizeof(int) * 4));
+            global.Size.Should().Be(sizeof(int) * 4);
         }
         finally
         {
@@ -30,7 +30,7 @@ public unsafe class HGLOBALTests
     {
         HGLOBAL global = default;
 
-        global.Size.Should().Be((nuint)0);
+        global.Size.Should().Be(0);
     }
 
     [TestMethod]
